Re-organize common module
