+++ tower-oauth2-resource-server/tests/common/jwks.rs
@@ -1,0 +1,33 @@
+use serde::Serialize;
+
+use crate::common::rsa::RsaKey;
+
+#[derive(Serialize)]
+pub struct Jwks {
+    keys: Vec<Jwk>,
+}
+
+#[derive(Serialize)]
+struct Jwk {
+    kty: String,
+    use_: String,
+    alg: String,
+    kid: String,
+    n: String,
+    e: String,
+}
+
+pub fn build_jwks(keys: &[(&str, &RsaKey)]) -> Jwks {
+    let keys = keys
+        .iter()
+        .map(|(kid, pub_key)| Jwk {
+            kty: "RSA".to_string(),
+            use_: "sig".to_string(),
+            alg: "RS256".to_string(),
+            kid: kid.to_string(),
+            n: pub_key.modulus.to_string(),
+            e: pub_key.exponent.to_string(),
+        })
+        .collect::<Vec<_>>();
+    Jwks { keys }
+}
