+++ tower-oauth2-resource-server/tests/common/oidc.rs
@@ -1,0 +1,7 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+pub struct OpenIdConfig {
+    pub issuer: String,
+    pub jwks_uri: String,
+}
